+++ src/main/java/com/project/webshopproject/ask/entity/Status.java
@@ -1,0 +1,5 @@
+package com.project.webshopproject.ask.entity;
+
+public enum Status {
+    WAITING, ANSWERED
+}
