--- src/main/java/com/project/webshopproject/model/Status.java
@@ -1,5 +1,0 @@
-package com.project.webshopproject.model;
-
-public enum Status {
-    WAITING, ANSWERED
-}
